move reviews functionality to new page
